+++ src/uniform_buffer.hpp
@@ -31,7 +31,7 @@ public:
 
     [[nodiscard]] bool operator<(const UniformBuffer& other) const;
 
-    [[nodiscard]] uint64_t handle() const;
+    [[nodiscard]] size_t handle() const;
 
     [[nodiscard]] bool is_valid() const;
 
@@ -52,7 +52,7 @@ public:
 private:
     bool m_valid = false;
     Renderer* m_renderer;
-    uint64_t m_handle;
+    size_t m_handle;
 };
 
 }
@@ -62,7 +62,7 @@ template <>
 struct hash<mve::UniformBuffer> {
     std::size_t operator()(const mve::UniformBuffer& uniform_buffer) const
     {
-        return hash<uint64_t>()(uniform_buffer.handle());
+        return hash<size_t>()(uniform_buffer.handle());
     }
 };
 }
